Rename single bit bloom filter gadget
