+++ src/gadgets.rs
@@ -1,4 +1,3 @@
-mod bloom_filter_old;
 mod bloom_filter;
 mod hash;
 mod response_accumulator;
