+++ src/gadgets/wnn.rs
@@ -8,9 +8,10 @@ use halo2_proofs::{
 use ndarray::{array, Array3};
 
 use crate::gadgets::{
-    bloom_filter_old::{
-        BloomFilterChip, BloomFilterChipConfig, BloomFilterConfig, BloomFilterInstructions,
+    bloom_filter::single_bit_bloom_filter::{
+        BloomFilterChip, BloomFilterChipConfig, BloomFilterConfig,
     },
+    bloom_filter::BloomFilterInstructions,
     hash::{HashChip, HashConfig, HashInstructions},
     response_accumulator::ResponseAccumulatorInstructions,
 };
